requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
nbformat==5.9.2
pyngrok==7.0.0
aiofiles==23.2.1
